Split CLI into tslint-cli and a CliRunner class
tslint-cli.ts takes in arguments using optimist and converts them to a
friendlier object to pass to a CliRunner. The rest of the logic that
used to be in tslint-cli.ts is now in the CliRunner.

Notes:
* I really wanted to make `CliRunner::run` return a `Promise<number>`,
but didn't want to add new dependencies or change the output target.
* I tried to stick to the existing code style for the cut-and-pasted
code. If you'd like it changed I'm up for that too.
* I'm not confident in `CliRunner` as a name. Alternate suggestions,
anyone?